refactor: Refactor all other installation tasks
