+++ WPILibInstaller-Avalonia/InstallTasks/Observer.cs
